backend/pyproject.toml
[project]
name = "glafrica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "asgiref==3.10.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "cloudinary==1.44.1",
    "colorama==0.4.6",
    "distro==1.9.0",
    "dj-database-url==3.0.1",
    "django==5.2.7",
    "django-cloudinary-storage==0.3.0",
    "django-cors-headers==4.9.0",
    "django-filter>=25.2",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jiter==0.11.1",
    "openai==2.6.1",
    "pgvector>=0.4.2",
    "pillow==12.0.0",
    "psycopg2-binary==2.9.11",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pyjwt==2.10.1",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
]
